pyproject.toml
[tool.poetry]
name = "writer"
version = "0.1.0"
description = ""
authors = ["Adrián Arroyo Calle <adrian.arroyocalle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
numpy = "^1.20.2"
scikit-learn = "^0.24.1"
joblib = "^1.0.1"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
